Refactoring logic for constructing markov automata - (Task #280)
---
Task #280: Improve time minimization of recovery automata
